Add files via upload
2023.11.23 iMON SPI 1st comm update

+++ ADS1115.h
@@ -68,6 +68,6 @@ void I2C0_IRQHandler(void);
 void I2C1_IRQHandler(void);
 void Delay(uint32_t delayCnt);
 
-int get_PinValue();
+uint8_t get_PinValue();
